secondcortex-backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
openai>=1.12.0
azure-identity>=1.15.0
chromadb>=0.4.20
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.27.0
gunicorn>=22.0.0
PyJWT>=2.8.0
mcp>=1.2.0
azure-ai-documentintelligence>=1.0.0b4
